refactor(api): Move calibration related functions out of the protocol api
