+++ api/src/opentrons/calibration_storage/types.py
@@ -1,0 +1,114 @@
+import typing
+from typing_extensions import TypedDict
+from dataclasses import dataclass
+from datetime import datetime
+
+
+CalibrationID = typing.NewType('CalibrationID', str)
+
+
+class TipLengthCalNotFound(Exception):
+    pass
+
+
+@dataclass
+class UriDetails:
+    namespace: str
+    load_name: str
+    version: int
+
+
+@dataclass
+class OffsetData:
+    """
+    Class to categorize the shape of a
+    given calibration data.
+    """
+    value: typing.List[float]
+    last_modified: typing.Optional[str]
+
+
+@dataclass
+class TipLengthData:
+    """
+    Class to categorize the shape of a
+    given calibration data.
+    """
+    value: typing.Optional[float] = None
+    last_modified: typing.Optional[str] = None
+
+
+@dataclass
+class ParentOptions:
+    """
+    Class to store whether a labware calibration has
+    a module, as well the original parent (slot).
+    As of now, the slot is not saved in association
+    with labware calibrations.
+    """
+    slot: str
+    module: str = ''
+
+
+@dataclass
+class CalibrationTypes:
+    """
+    Class to categorize what calibration
+    data might be stored for a labware.
+    """
+    offset: OffsetData
+    tip_length: TipLengthData
+
+
+@dataclass
+class CalibrationInformation:
+    """
+    Class to store important calibration
+    info for labware.
+    """
+    calibration: CalibrationTypes
+    parent: ParentOptions
+    labware_id: str
+    uri: str
+
+
+class TipLengthCalibration(TypedDict):
+    tipLength: float
+    lastModified: datetime
+
+
+class ModuleDict(TypedDict):
+    parent: str
+    fullParent: str
+
+
+class CalibrationIndexDict(TypedDict):
+    """
+    The dict that is returned from
+    the index.json file.
+    """
+    uri: str
+    slot: str
+    module: ModuleDict
+
+
+class OffsetDict(TypedDict):
+    offset: typing.List[float]
+    lastModified: str
+
+
+class TipLengthDict(TypedDict):
+    length: float
+    lastModified: str
+
+
+class CalibrationDict(TypedDict):
+    """
+    The dict that is returned from a labware
+    offset file.
+    """
+    default: OffsetDict
+    tipLength: TipLengthDict
+
+
+PipTipLengthCalibration = typing.Dict[str, TipLengthCalibration]
