+++ api/src/opentrons/protocol_api/constants.py
@@ -5,5 +5,4 @@ from opentrons.config import get_opentrons_path
 OPENTRONS_NAMESPACE = 'opentrons'
 CUSTOM_NAMESPACE = 'custom_beta'
 STANDARD_DEFS_PATH = Path("labware/definitions/2")
-OFFSETS_PATH = get_opentrons_path('labware_calibration_offsets_dir_v2')
 USER_DEFS_PATH = get_opentrons_path('labware_user_definitions_dir_v2')
