+++ api/src/opentrons/protocol_api/dev_types.py
@@ -1,5 +1,4 @@
-from datetime import datetime
-from typing import Callable, Dict, TYPE_CHECKING, List
+from typing import Callable, Dict, TYPE_CHECKING
 
 from typing_extensions import Protocol, TypedDict
 
@@ -130,46 +129,3 @@ JsonV4ThermocyclerDispatch = TypedDict(
             ['ThermocyclerContext', 'ModuleIDParams'], None]
     }
 )
-
-
-# TODO: AA 2020-06-10 move these out of protocol_api
-class TipLengthCalibration(TypedDict):
-    tipLength: float
-    lastModified: datetime
-
-
-class ModuleDict(TypedDict):
-    parent: str
-    fullParent: str
-
-
-class CalibrationIndexDict(TypedDict):
-    """
-    The dict that is returned from
-    the index.json file.
-    """
-    uri: str
-    slot: str
-    module: ModuleDict
-
-
-class OffsetDict(TypedDict):
-    offset: List[float]
-    lastModified: str
-
-
-class TipLengthDict(TypedDict):
-    length: float
-    lastModified: str
-
-
-class CalibrationDict(TypedDict):
-    """
-    The dict that is returned from a labware
-    offset file.
-    """
-    default: OffsetDict
-    tipLength: TipLengthDict
-
-
-PipTipLengthCalibration = Dict[str, TipLengthCalibration]
